requirements.txt
av==10.0.0
opencv-python-headless == 4.9.0.80
streamlit-webrtc==0.47.1
tensorflow==2.14.0
